float to double

+++ examples/ex1mod4p.cpp
@@ -77,9 +77,9 @@ class MassMatrix1 : public MatrixCoefficient
 {
    private:
       GridFunction &phi; // vector-valued GridFunction
-	  float alpha;
+	  double alpha;
    public:
-      MassMatrix1(GridFunction &phi_, float alpha_) : MatrixCoefficient(phi_.FESpace()->GetVDim()), phi(phi_), alpha(alpha_) {}
+      MassMatrix1(GridFunction &phi_, double alpha_) : MatrixCoefficient(phi_.FESpace()->GetVDim()), phi(phi_), alpha(alpha_) {}
 
    virtual void Eval(DenseMatrix &M, ElementTransformation &T, const IntegrationPoint &ip)
    {
@@ -114,9 +114,9 @@ class LambdaDivPart : public VectorCoefficient
       GridFunction &phi; // vector-valued GridFunction
       GridFunction &gradX; // vector-valued GridFunction
       GridFunction &gradY; // vector-valued GridFunction
-	  float alpha;
+	  double alpha;
    public:
-      LambdaDivPart(GridFunction &phi_, GridFunction &gradX_, GridFunction &gradY_, float alpha_) : 
+      LambdaDivPart(GridFunction &phi_, GridFunction &gradX_, GridFunction &gradY_, double alpha_) : 
            VectorCoefficient(phi_.FESpace()->GetVDim()), phi(phi_), gradX(gradX_), gradY(gradY_), alpha(alpha_) {}
 
    virtual void Eval(Vector &V, ElementTransformation &T, const IntegrationPoint &ip)
@@ -368,7 +368,7 @@ int main(int argc, char *argv[])
     [](const Vector &x, Vector &y) { y = 0.0; });   
    VectorFunctionCoefficient bump(pmesh.Dimension(),
     [](const Vector &x, Vector &y) { 
-	  float width = 0.2;
+	  double width = 0.2;
 	  if(abs(x[0]) < width && abs(x[1]) < width){
 		  y[0] = (width-abs(x[0]))*(width-abs(x[0]))*0.01;
 		  y[1] = (width-abs(x[1]))*(width-abs(x[1]))*0.01;
@@ -378,7 +378,6 @@ int main(int argc, char *argv[])
 	});
    VectorFunctionCoefficient shock(pmesh.Dimension(),
     [](const Vector &x, Vector &y) { 
-	  //float width = 0.2;
 	  if(x[0] < 0){
 		  y[0] = 1.0;
 		  y[1] = 0.0;
@@ -424,7 +423,7 @@ int main(int argc, char *argv[])
       a.EnableSparseMatrixSorting(Device::IsEnabled());
    }
    
-   float alpha = 0.1; 
+   double alpha = 0.1; 
    a.AddDomainIntegrator(new MassIntegrator);
